Added questions being asked

+++ headers/game.h
@@ -27,7 +27,7 @@ private:
   Enemy _enemy;
   Sprite _sprite;
 
-  float _score;
+  int _score;
 
   SDL_Texture* _tileset;
